Added list of ontology sources
Unable to locate for the dbpedia prefixes (not surprising given the quantity of terms).  The "xml" namespace refers to an outdated vocab and I don't think we use it.  That would also impact the xsd namespace.  The issue of using "schema" as a namespace is one on the ontology improvements list, and the "lemon" website is temporarily down.