Added doi to turtle
Note: I recognize that the presence of bibo:doi could make gcis:hasURL redundant.  However, this isn't trivial for the NCA3 since the url directs one to the web version of that document, while the other to the download page.  (Fine with me, of course).  Of course this won't hold true for all "chapters" in GCIS.  If including both "doi" and "hasURL" is problematic, let me know and I'll create an issue in gethub.