Added DB triggers for copy_exterm_inserts_to_term and standard auditing, and address comments from pull request.
NOTE: as documented in
http://dba.stackexchange.com/questions/123691/output-from-pg-dump-schema-only-has-inconsistent-order
the triggers for 'model' and 'model_run' tables may be in differing orders in
the schema verification during dbfakeinstall, and cause a warning.  This does
NOT actually reflect an issue with the schema integrety, but appears to be a
pgdump bug using arbitrary ordering for triggers when dumping the schema.