Add count_only param for /search
Principle changes are in 'keyword' and 'dbgrep', with additional changes
to accomodate count_only skeletal Rose::DB::Objects not supporting the
stringify method (nor really needing to).

Of note, dbgrep is a good candidate for refactoring, to avoid having to
duplicate changes in the DB/Mixin/Object code for Chapter, Figure, Finding,
Report, and Table.