Fix JSON thumbnail_href value
This was previously an invalid value being returned in JSON.
I see no reason why it shouldn't just use the same logic that the
html pages use for getting the thumbnail.  If Tuba.conf has the line
   with_color_images : 1
then the thumbnails will be fetched in color.

Bug was: (eg, for report/usgs-profpaper-1775.json)
Abbreviated JSON result was
{ files [ { thumbnail_href: http://ec2-52-38-26-42.us-west-2.compute.amazonaws.com:8080/assets/55/cf/0fa00f85af295ddf5d2346d7fb8d/.thumb-f290591d-9065-4da3-a3e4-29f64c60739f.png } ] }

It now is
{ files [ { thumbnail_href: http://data.globalchange.gov/assets/55/cf/0fa00f85af295ddf5d2346d7fb8d/.thumb-f290591d-9065-4da3-a3e4-29f64c60739f.png } ] }