Fix Term recursion to only detect within a branch
Using the same hashref for the entire tree allowed branches to "see"
nodes from other branches.  It's a kind of loop, but not a recursive one.
Eg, the following is not recursive, but node Z would be pruned as a child
of Y, because the C branch already listed it:
    A - B - C
     \       \
      X - Y - Z - Q

That's not the behavior we want.  ABCZQ and AXYZQ now will both exist as
full sub-branches.