Make ::Object::Term::uri like Tuba::DB::Object
The uri method in Tuba::DB::Object::Term is overridden really only
to use a unique key instead of the primary key of the table for the
uri.  The code was originally a direct copy-and-paste, to prototype
the feasibility of passing a unique-key parameter, which would
eliminate the need for overriding the method.  Due to changes in
url_for behavior as of Mojolicious 7.x, additional debugging is
being added here, as was done in the superclass method.